+++ app/src/main/java/app/passwordstore/util/settings/PreferenceKeys.kt
@@ -22,6 +22,7 @@ object PreferenceKeys {
   const val COPY_ON_DECRYPT = "copy_on_decrypt"
   const val ENABLE_DEBUG_LOGGING = "enable_debug_logging"
   const val EXPORT_PASSWORDS = "export_passwords"
+  const val IMPORT_PASSWORDS = "import_passwords"
   const val FILTER_RECURSIVELY = "filter_recursively"
   const val GENERAL_SHOW_TIME = "general_show_time"
   const val GIT_CONFIG = "git_config"